# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Laravel API",
        "description": "API documentation for my Laravel 12 project.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8085/api",
            "description": "Main API server"
        }
    ],
    "paths": {
        "/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get a paginated list of articles",
                "description": "Retrieve articles with optional filters for category, author, source, publish date range, and free-text search across title, description, content, and keywords.",
                "operationId": "89181b74a8aa4d580750ede6aa32f92a",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by category",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "World News"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "description": "Filter by author",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Benjamen"
                        }
                    },
                    {
                        "name": "source_id",
                        "in": "query",
                        "description": "Filter by source identifier",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    },
                    {
                        "name": "publish_date",
                        "in": "query",
                        "description": "Date range filter in the format 'YYYY-MM-DD,YYYY-MM-DD'",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "2025-01-01,2025-01-31"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Full-text search across title, description, content, and keywords",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "climate change"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of results per page (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated list of articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ArticlesResource"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get a single article",
                "description": "Fetch the details of a specific article by its identifier.",
                "operationId": "b3e529eab04efb5d305b07e2e58d9869",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Article details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/ArticlesResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found"
                    }
                }
            }
        },
        "/sources": {
            "get": {
                "tags": [
                    "Sources"
                ],
                "summary": "Get a list of all sources",
                "description": "Retrieve a paginated list of news sources with optional filtering and sorting",
                "operationId": "21bca661d3026520af28d6595d6178fa",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page (default: 15, max: 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "key": {
                                                        "type": "string",
                                                        "example": "bbc-news"
                                                    },
                                                    "label": {
                                                        "type": "string",
                                                        "example": "BBC News"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "path": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://api.example.com/api/articles"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 15,
                        "nullable": true
                    },
                    "total": {
                        "type": "integer",
                        "example": 150
                    }
                },
                "type": "object"
            },
            "ArticlesResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "slug": {
                        "type": "string",
                        "example": "breaking-news-update"
                    },
                    "provider_id": {
                        "type": "string",
                        "example": "bbc-world-123",
                        "nullable": true
                    },
                    "provider": {
                        "type": "string",
                        "example": "bbc-news"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "content": {
                        "type": "string"
                    },
                    "image_url": {
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "source_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "author": {
                        "type": "string",
                        "nullable": true
                    },
                    "link": {
                        "type": "string",
                        "format": "uri"
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "category": {
                        "type": "string",
                        "nullable": true
                    },
                    "language": {
                        "type": "string",
                        "nullable": true
                    },
                    "keywords": {
                        "type": "string",
                        "nullable": true
                    },
                    "source": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/SourceResource"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "SourceResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3
                    },
                    "key": {
                        "type": "string",
                        "example": "cnn"
                    },
                    "label": {
                        "type": "string",
                        "example": "CNN"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Articles",
            "description": "Articles"
        },
        {
            "name": "Sources",
            "description": "Sources"
        }
    ]
}